requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
joblib==1.2.0
python-dotenv==0.21.0
scikit-learn==1.3.0
google-generativeai==0.8.3
google-api-core==2.23.0
pydantic==1.10.8
pandas==2.0.3
tldextract==3.2.0
